viewmodel 문제원인파악
