+++ app/src/main/java/com/kang/floapp/view/main/frag/FragPlaylist.java
@@ -35,6 +35,8 @@ public class FragPlaylist extends Fragment {
         mp = mainActivity.mp;
         mainViewModel = mainActivity.mainViewModel;
 
+
+        //아하!!! 이거 먼저 띄우고 해야 view model이 작동하는구나...어쩌지....
         rvPlayList = view.findViewById(R.id.rv_play_list);
         LinearLayoutManager layoutManager = new LinearLayoutManager(getActivity());
         rvPlayList.setLayoutManager(layoutManager);
